fix: prefetch 적용
