+++ src/app/group/[groupId]/page.tsx
@@ -1,13 +1,15 @@
 import { getGroupInfo, getGroupMemberList } from "@/app/api/groups";
-import { getTopRanking } from "@/app/api/groups/ranking";
+import { getAllRanking, getTopRanking } from "@/app/api/groups/ranking";
 import { getDeadlineReachedProblems } from "@/app/api/problems";
 import { listSectionStyle, titleStyle } from "@/app/group/[groupId]/page.css";
 import Sidebar from "@/common/component/Sidebar";
 import ProblemList from "@/shared/component/ProblemList";
+import { prefetchQuery } from "@/shared/util/prefetch";
 import { sidebarWrapper } from "@/styles/shared.css";
 import GroupSidebar from "@/view/group/dashboard/GroupSidebar";
 import NoticeBanner from "@/view/group/dashboard/NoticeBanner";
 import Ranking from "@/view/group/dashboard/Ranking";
+import { HydrationBoundary } from "@tanstack/react-query";
 
 const GroupDashboardPage = async ({
   params: { groupId },
@@ -25,6 +27,12 @@ const GroupDashboardPage = async ({
       deadlineReachedData,
     ]);
 
+  const firstPage = 1;
+  const dehydratedState = await prefetchQuery({
+    queryKey: ["ranking", +groupId, firstPage],
+    queryFn: () => getAllRanking(+groupId, 0),
+  });
+
   return (
     <main className={sidebarWrapper}>
       <Sidebar>
@@ -32,7 +40,9 @@ const GroupDashboardPage = async ({
       </Sidebar>
       <div className={listSectionStyle}>
         <NoticeBanner />
-        <Ranking rankingData={rankingInfo} />
+        <HydrationBoundary state={dehydratedState}>
+          <Ranking rankingData={rankingInfo} />
+        </HydrationBoundary>
         <h2 className={titleStyle}>풀어야 할 문제</h2>
         <section>
           <ProblemList.Header />
