+++ src/app/provider.tsx
@@ -2,12 +2,8 @@
 
 import ToastProvider from "@/common/component/Toast";
 import JotaiProvider from "@/shared/component/Provider";
+import QueryProvider from "@/shared/component/QueryProvider";
 import RefreshTokenExpireTime from "@/shared/component/RefreshTokenExpireTime";
-import {
-  QueryClient,
-  QueryClientProvider,
-  isServer,
-} from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { useSession } from "next-auth/react";
 import dynamic from "next/dynamic";
@@ -24,36 +20,11 @@ type ProvidersProps = {
   children: ReactNode;
 };
 
-const createQueryClient = () => {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: 60 * 1000,
-      },
-    },
-  });
-};
-
-let browserQueryClient: QueryClient | undefined = undefined;
-
-const getQueryClient = () => {
-  if (isServer) {
-    return createQueryClient();
-  }
-  if (!browserQueryClient) {
-    browserQueryClient = createQueryClient();
-  }
-
-  return browserQueryClient;
-};
-
 const Providers = ({ children }: ProvidersProps) => {
-  const queryClient = getQueryClient();
-
   const { data: session, update } = useSession();
 
   return (
-    <QueryClientProvider client={queryClient}>
+    <QueryProvider>
       <RefreshTokenExpireTime session={session} update={update} />
       <BrowserProvider>
         <JotaiProvider>
@@ -62,7 +33,7 @@ const Providers = ({ children }: ProvidersProps) => {
         </JotaiProvider>
       </BrowserProvider>
       <ReactQueryDevtools initialIsOpen={false} />
-    </QueryClientProvider>
+    </QueryProvider>
   );
 };
