+++ src/shared/component/QueryProvider/index.tsx
@@ -3,12 +3,24 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { useState } from "react";
 
-export default function QueryProvider({
+const createQueryClient = () => {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: 60 * 1000,
+      },
+    },
+  });
+};
+
+const QueryProvider = ({
   children,
 }: {
   children: React.ReactNode;
-}) {
-  const [client] = useState(new QueryClient());
+}) => {
+  const [client] = useState(createQueryClient());
 
   return <QueryClientProvider client={client}>{children}</QueryClientProvider>;
-}
+};
+
+export default QueryProvider;
